+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Center on selected elements on zoom in, zoom out and zoom one to one.
 * Fixed bugs:
   * failure parsing demo file with "Plus number must be positive" parse error.
+  * parsing/writing '+?' and '-?' values failure for point height -value options.
 * Infrastructure maintenance:
 
 ## 0.2.29 - 2025-12-27 - The Bom Jesus da Lapa release